# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Pulse API",
    "description": "REST API for Pulse",
    "contact": {
      "name": "Brian Mendoza",
      "email": "bcmendoza@gmail.com"
    },
    "license": {
      "name": "License",
      "url": "https://github.com/bcmendoza/pulse/blob/master/LICENSE"
    }
  },
  "definitions": {
    "Stream": {
      "type": "object",
      "properties": {
        "unitType": {
          "type": "string",
          "example": "BPM"
        },
        "history": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Pulse"
          }
        },
        "lower": {
          "type": "number",
          "format": "float",
          "example": 33.3
        },
        "upper": {
          "type": "number",
          "format": "float",
          "example": 82.1
        }
      }
    },
    "Pulse": {
      "type": "object",
      "properties": {
        "score": {
          "type": "number",
          "format": "float",
          "example": 27.36
        },
        "timestamp": {
          "type": "string",
          "example": "2017-07-21T17:32:28Z",
          "format": "date-time"
        },
        "rating": {
          "type": "integer",
          "example": 2
        }
      }
    },
    "Hospital": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "hospital"
        },
        "children": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/Department"
          }
        },
        "stream": {
          "$ref": "#/definitions/Stream"
        }
      }
    },
    "Department": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "department"
        },
        "name": {
          "type": "string",
          "example": "ICU"
        },
        "children": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/Patient"
          }
        },
        "stream": {
          "$ref": "#/definitions/Stream"
        }
      }
    },
    "Patient": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "patient"
        },
        "name": {
          "type": "string",
          "example": "39fd6a79-1791-4117-9d14-9a7b2118f930"
        },
        "children": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/Metric"
          }
        },
        "stream": {
          "$ref": "#/definitions/Stream"
        }
      }
    },
    "Metric": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "metric"
        },
        "name": {
          "type": "string",
          "example": "Heart Rate"
        },
        "stream": {
          "$ref": "#/definitions/Stream"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "example": 404
        },
        "message": {
          "type": "string",
          "example": "Not Found"
        }
      }
    }
  },
  "paths": {
    "/streams": {
      "get": {
        "description": "Gets a streams object of the entire hospital",
        "tags": ["Streams"],
        "produces": ["application/json"],
        "operationId": "streams",
        "responses": {
          "200": {
            "description": "Streams Response",
            "schema": {
              "$ref": "#/definitions/Hospital"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/departments": {
      "post": {
        "description": "Adds a new department to the hospital",
        "tags": ["Departments"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "operationId": "departments",
        "parameters": [
          {
            "name": "AddDepartmentRequest",
            "in": "body",
            "description": "Add a department",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "department": {
                  "type": "string",
                  "example": "ICU"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "added": {
                  "type": "string",
                  "example": "NICU"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/patients": {
      "post": {
        "description": "Adds a new patient to a department. Patient ID autogenerated",
        "tags": ["Patients"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "operationId": "patients",
        "parameters": [
          {
            "name": "AddPatientRequest",
            "in": "body",
            "description": "Add a patient",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "department": {
                  "type": "string",
                  "example": "ICU"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "added": {
                  "type": "string",
                  "example": "NICU"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/metrics": {
      "post": {
        "description": "Adds a new metric to a patient, entire department, or entire hospital",
        "tags": ["Metrics"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "operationId": "metrics",
        "parameters": [
          {
            "name": "AddMetricRequest",
            "in": "body",
            "description": "Add a metric. To apply to all patients, leave 'patient' and 'department' keys empty. To apply to a department's patients, leave 'patient' empty. Otherwise, to apply to a patient, all fields are required.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "department": {
                  "type": "string",
                  "example": "ICU"
                },
                "patient": {
                  "type": "string",
                  "example": "39fd6a79-1791-4117-9d14-9a7b2118f930"
                },
                "metric": {
                  "type": "string",
                  "example": "Heart Rate"
                },
                "unitType": {
                  "type": "string",
                  "example": "BPM"
                },
                "lower": {
                  "type": "number",
                  "format": "float",
                  "example": 51.01
                },
                "upper": {
                  "type": "number",
                  "format": "float",
                  "example": 120.09
                }
              },
              "required": ["metric", "unitType", "lower", "upper"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "added": {
                  "type": "string",
                  "example": "NICU"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/pulses": {
      "post": {
        "description": "Adds a metric pulse, updating all listening streams",
        "tags": ["Pulses"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "operationId": "pulses",
        "parameters": [
          {
            "name": "AddMetricPulseRequest",
            "in": "body",
            "description": "Add a metric pulse",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "department": {
                  "type": "string",
                  "example": "ICU"
                },
                "patient": {
                  "type": "string",
                  "example": "patient-39fd6a79-1791-4117-9d14-9a7b2118f930"
                },
                "metric": {
                  "type": "string",
                  "example": "Heart Rate"
                },
                "value": {
                  "type": "number",
                  "format": "float",
                  "example": 93.26
                }
              },
              "required": ["department", "patient", "metric", "value"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "added": {
                  "type": "number",
                  "format": "float",
                  "example": 93.26
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  }
}